Move functions into utility files
